refactor: extract into view objects
Not sure how useful this change is, but I like it better without
`/app/helpers` so let's try.
