+++ app/view_objects/view_object.rb
@@ -1,0 +1,4 @@
+class ViewObject
+  include Rails.application.routes.url_helpers
+  include ActionView::Helpers
+end
