+++ app/views/catalog/_show/_editor_buttons.haml
@@ -1,6 +1,6 @@
 .right
   -if user_is_editor?
-    =link_to_edit_taxon taxon
+    =editors_taxon_view_object.link_to_edit_taxon
   =link_to "History", taxon_history_path(taxon), class: "btn-normal"
   =link_to "What Links Here", taxon_what_links_here_path(taxon), class: "btn-normal"
 
@@ -8,15 +8,15 @@
     .btn-normal.dropdown.button{data: { toggle: "more-dropdown"}, type: "button"} More
     #more-dropdown.dropdown-pane{data: { dropdown: 'true', hover: 'true', hover_pane: 'true' }}
       %ul.no-bullet
-        %li=add_child_button taxon
-        %li=add_tribe_button taxon
-        %li=add_subgenus_button taxon
+        %li=editors_taxon_view_object.add_child_button
+        %li=editors_taxon_view_object.add_tribe_button
+        %li=editors_taxon_view_object.add_subgenus_button
 
         -if taxon.is_a?(Species)
           %li=link_to "Set subgenus", taxa_set_subgenus_path(taxon), class: "btn-normal"
 
-        %li=elevate_to_species_button taxon
-        %li=convert_to_subspecies_button taxon
+        %li=editors_taxon_view_object.elevate_to_species_button
+        %li=editors_taxon_view_object.convert_to_subspecies_button
 
         -if taxon.policy.show_create_combination_button?
           %li=link_to "Create combination#{' (blocked)' unless taxon.policy.allow_create_combination? }", new_taxa_create_combination_path(taxon), class: "btn-normal"
@@ -30,7 +30,7 @@
         -if taxon.policy.allow_force_change_parent?
           %li=link_to "Force parent change...", taxa_force_parent_change_path(taxon), class: "btn-warning"
 
-        %li=delete_unreferenced_taxon_button taxon
+        %li=editors_taxon_view_object.delete_unreferenced_taxon_button
         %li=link_to "Move items", new_taxa_move_items_path(taxon), class: "btn-normal"
         %li=link_to "Add history item", new_taxa_taxon_history_item_path(taxon), class: "btn-normal"
         %li=link_to "Show children", taxa_children_path(taxon), class: "btn-normal"
