+++ app/views/catalog/_show/_editors_sidebar.haml
@@ -43,13 +43,13 @@
           =taxon.protonym.decorate.link_to_protonym
       =render 'shared/failed_soft_validations', soft_validations: taxon.protonym.soft_validations
 
--if editor_taxon_view_object.taxa_with_same_name.present?
+-if editors_taxon_view_object.taxa_with_same_name.present?
   .additional-editor-data.callout.no-border-callout.logged-in-only-background
     %h6
       Taxa with same name
       =logged_in_only_tooltip_icon
     %ul.compact-list.small-margin-bottom
-      -editor_taxon_view_object.taxa_with_same_name.includes(:name, protonym: [{ authorship: :reference }]).each do |homonym|
+      -editors_taxon_view_object.taxa_with_same_name.includes(:name, protonym: [{ authorship: :reference }]).each do |homonym|
         %li
           =homonym.link_to_taxon
           =homonym.author_citation
